move chan enums
