+++ server/models/network.ts
@@ -1,7 +1,7 @@
 import _ from "lodash";
 import {v4 as uuidv4} from "uuid";
 import IrcFramework, {Client as IRCClient} from "irc-framework";
-import Chan, {ChanConfig, Channel, ChanType} from "./chan";
+import Chan, {ChanConfig, Channel} from "./chan";
 import Msg from "./msg";
 import Prefix from "./prefix";
 import Helper, {Hostmask} from "../helper";
@@ -10,6 +10,7 @@ import STSPolicies from "../plugins/sts";
 import ClientCertificate, {ClientCertificateType} from "../plugins/clientCertificate";
 import Client from "../client";
 import {MessageType} from "../../shared/types/msg";
+import {ChanType} from "../../shared/types/chan";
 
 /**
  * List of keys which should be sent to the client by default.
