+++ server/plugins/inputs/ban.ts
@@ -1,7 +1,7 @@
-import {ChanType} from "../../models/chan";
 import Msg from "../../models/msg";
 import {PluginInputHandler} from "./index";
 import {MessageType} from "../../../shared/types/msg";
+import {ChanType} from "../../../shared/types/chan";
 
 const commands = ["ban", "unban", "banlist", "kickban"];
