+++ server/plugins/inputs/ignore.ts
@@ -2,8 +2,8 @@ import Msg from "../../models/msg";
 import Helper from "../../helper";
 import {PluginInputHandler} from "./index";
 import {IgnoreListItem} from "../../models/network";
-import {ChanType, SpecialChanType} from "../../models/chan";
 import {MessageType} from "../../../shared/types/msg";
+import {ChanType, SpecialChanType} from "../../../shared/types/chan";
 
 const commands = ["ignore", "unignore", "ignorelist"];
