+++ server/plugins/irc-events/away.ts
@@ -1,7 +1,7 @@
 import {IrcEventHandler} from "../../client";
-import {ChanType} from "../../models/chan";
 import Msg from "../../models/msg";
 import {MessageType} from "../../../shared/types/msg";
+import {ChanType} from "../../../shared/types/chan";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
