+++ server/plugins/irc-events/join.ts
@@ -1,8 +1,8 @@
 import Msg from "../../models/msg";
 import User from "../../models/user";
 import type {IrcEventHandler} from "../../client";
-import {ChanState} from "../../models/chan";
 import {MessageType} from "../../../shared/types/msg";
+import {ChanState} from "../../../shared/types/chan";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
