+++ server/plugins/irc-events/list.ts
@@ -1,6 +1,7 @@
 import {IrcEventHandler} from "../../client";
 
-import Chan, {ChanType, SpecialChanType} from "../../models/chan";
+import Chan from "../../models/chan";
+import {ChanType, SpecialChanType} from "../../../shared/types/chan";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
