+++ server/plugins/irc-events/modelist.ts
@@ -1,8 +1,8 @@
 import {IrcEventHandler} from "../../client";
-import {SpecialChanType, ChanType} from "../../models/chan";
 
 import Msg from "../../models/msg";
 import {MessageType} from "../../../shared/types/msg";
+import {SpecialChanType, ChanType} from "../../../shared/types/chan";
 
 export default <IrcEventHandler>function (irc, network) {
 	const client = this;
